star using tab bar navigation at top level
