+++ apps/expo/src/app/(signedin)/(main)/home.tsx
@@ -3,7 +3,7 @@ import { useRouter } from "expo-router";
 
 import { LatestReviewSection } from "~/components/previous-review/latest-review-section";
 import { ThemeableView } from "~/components/themeable/themable-view";
-import { SearchInput } from "../../components/search-input";
+import { SearchInput } from "../../../components/search-input";
 
 export default function HomeScreen() {
   const [searchQuery, setSearch] = React.useState("");
