+++ apps/expo/src/app/(signedin)/(main)/review/detail/[id].tsx
