+++ apps/expo/src/app/(signedin)/(main)/review/new.tsx
@@ -13,9 +13,9 @@ import { TagInput } from "~/components/tag-input/tag-input";
 import { ThemeableText } from "~/components/themeable/themable-text";
 import { useCreateReview } from "~/hooks/queries/use-create-review";
 import { usePrimaryColor } from "~/hooks/use-primary-color";
-import { StyledButton } from "../../../components/button";
-import { StyledTextInput } from "../../../components/inputs/styled-text-input";
-import { StarRating } from "../../../components/star-rating";
+import { StyledButton } from "../../../../components/button";
+import { StyledTextInput } from "../../../../components/inputs/styled-text-input";
+import { StarRating } from "../../../../components/star-rating";
 
 interface ReviewInput {
   rating: number;
